bot + gitignore

+++ src/main/java/fr/clbd/fire/service/FireService.java
@@ -16,7 +16,7 @@ public class FireService {
     }
 
     public int getDistanceBetweenCoord(double lat1, double lon1, double lat2, double lon2) {
-        return RequestsUtils.getDistanceBetweenCoord(lat1, lon1, lat2, lon2);
+        return (int) RequestsUtils.getDistanceBetweenCoord(lat1, lon1, lat2, lon2);
     }
 }
